+++ src/unbroadcast.jl
@@ -1,0 +1,37 @@
+"""
+    unbroadcast(x,dx)
+
+Bring dx to x's size via unbroadcasting (reduction). This is needed
+when defining gradients of multi-argument broadcasting functions where
+the arguments and the result may be of different sizes.
+
+"""
+function unbroadcast(x, dx)
+    if size(x)==size(dx)
+        return dx
+    elseif isa(value(x),Number)
+        return sum(dx)
+    elseif isa(value(dx),Number)
+        return fill!(similar(value(x)),dx)
+    else
+        d = []
+        for i=1:ndims(dx)
+            size(x,i) == size(dx,i) > 1 && continue
+            size(x,i) != 1 && throw(DimensionMismatch())
+            push!(d,i)
+        end
+        length(d)==1 && (d=d[1])
+        return reshape(sum(dx, dims=d), size(x))
+    end
+end
+
+# This fixes unbroadcast when x isa Broadcasted
+Base.size(x::Base.Broadcast.Broadcasted,i...)=length.(axes(x,i...))
+
+# Broadcasting dimensions:
+# y = f(x...) where f is a broadcasting operation.
+# size(y) = broadcast_shape(x...)
+# ndims(y) = max ndims(x)
+# size(y,i) = max size(x,i)
+# size(x,i) = 1 or size(y,i) for all x and i<=ndims(x)
+# if ndims(x) < ndims(y) the extra dimensions of x are treated as 1